refactor: rename new components to replace old and move old to new dir
