+++ components/CustomButton.tsx
@@ -1,13 +1,11 @@
-import { Pressable } from "react-native";
-import { Text } from "react-native";
-import globalStyle from "../style";
+import { Pressable, Text } from "react-native";
 
 interface ButtonProps {
   label: React.ReactNode | string;
   action: () => void;
 }
 
-const CustomButton = ({ label, action }: ButtonProps) => {
+export default function CustomButton({ label, action }: ButtonProps) {
   return (
     <Pressable
       className=""
@@ -15,17 +13,10 @@ const CustomButton = ({ label, action }: ButtonProps) => {
       hitSlop={{ bottom: 15, left: 15, right: 15, top: 15 }}
     >
       {typeof label === "string" ? (
-        <Text
-          style={globalStyle.fontStyle}
-          className="text-gray-400 capitalize tracking-wider"
-        >
-          {label}
-        </Text>
+        <Text className="text-gray-400 capitalize tracking-wider">{label}</Text>
       ) : (
         label
       )}
     </Pressable>
   );
-};
-
-export default CustomButton;
+}
