+++ components/Footer.tsx
@@ -1,14 +1,11 @@
 import { View, Text } from "react-native";
 import CustomButton from "./CustomButton";
-import globalStyle from "../style";
 
-const Footer = () => {
+export default function Footer() {
   return (
     <View className="px-4 py-6 flex-row justify-between">
       <View>
-        <Text style={globalStyle.fontStyle} className="text-gray-400">
-          5 items left
-        </Text>
+        <Text className="text-gray-400">5 items left</Text>
       </View>
       <View>
         <CustomButton
@@ -18,6 +15,4 @@ const Footer = () => {
       </View>
     </View>
   );
-};
-
-export default Footer;
+}
