+++ components/Input.tsx
@@ -1,34 +1,26 @@
+import { View, TextInput } from "react-native";
 import { useState } from "react";
-import { View, TextInput, StyleSheet } from "react-native";
 import CustomCheckBox from "./CustomCheckBox";
-import globalStyle from "../style";
 
-const Input = () => {
+export default function Input() {
   const [task, setNewTask] = useState("");
+
   return (
-    <View className="bg-element-violet h-14 rounded-md flex-row items-center overflow-hidden">
-      <View
-        style={{ flex: 0.2 }}
-        className="h-full items-center justify-center"
-      >
+    <View
+      className="flex-row rounded-md bg-element-violet overflow-hidden py-2"
+      style={{ width: "85%" }}
+    >
+      <View style={{ flex: 0.15 }} className="items-center justify-center p-2">
         <CustomCheckBox />
       </View>
       <TextInput
-        className="px-4 text-white h-full"
-        style={[{ flex: 0.8 }, globalStyle.fontStyle]}
-        placeholder="Create a new todo..."
+        className="p-2 text-white"
+        style={{ flex: 0.85 }}
         value={task}
         onChangeText={setNewTask}
         placeholderTextColor="#9ca3af"
+        placeholder="Create a new todo..."
       />
     </View>
   );
-};
-
-const styles = StyleSheet.create({
-  container: {
-    // height: 70,
-  },
-});
-
-export default Input;
+}
