+++ components/failsafe/ListItem.tsx
@@ -1,25 +1,30 @@
 import { View, Text } from "react-native";
-import CustomCheckBox from "./CustomCheckBoxCopy";
+import CustomCheckBox from "./failsafe/CustomCheckBox";
+import CustomButton from "./failsafe/CustomButton";
 import { AntDesign } from "@expo/vector-icons";
-import CustomButton from "./CustomButtonCopy";
+import globalStyle from "../style";
 
 interface ListItemProps {
   todo: string;
 }
 
-export default function ListItem({ todo }: ListItemProps) {
+const ListItem = ({ todo }: ListItemProps) => {
   return (
     <View className="py-6 px-4 flex-row justify-between items-center border-b border-gray-500">
       <View>
         <CustomCheckBox />
       </View>
-      <Text className="text-gray-300 w-3/4">{todo}</Text>
+      <Text className="text-gray-300 w-3/4" style={globalStyle.fontStyle}>
+        {todo}
+      </Text>
       <View>
         <CustomButton
           label={<AntDesign name="close" size={24} color="#6b7280" />}
-          action={() => alert("Remove Todo")}
+          action={() => alert("Close Button")}
         />
       </View>
     </View>
   );
-}
+};
+
+export default ListItem;
